server/requirements.txt
fastapi
uvicorn
python-dotenv



uvicorn app.main:app --reload;   